+++ entity/src/main/java/com/gestor/seguimiento/PlanSeccionTextoPK.java
@@ -19,7 +19,7 @@ public class PlanSeccionTextoPK implements Serializable {
 
     @Basic(optional = false)
     @Column(name = "codigo_establecimiento")
-    private short codigoEstablecimiento;
+    private int codigoEstablecimiento;
     @Basic(optional = false)
     @Column(name = "cod_titulo")
     private int codTitulo;
@@ -33,18 +33,18 @@ public class PlanSeccionTextoPK implements Serializable {
     public PlanSeccionTextoPK() {
     }
 
-    public PlanSeccionTextoPK(short codigoEstablecimiento, int codTitulo, int codSeccion, int codSeccionTexto) {
+    public PlanSeccionTextoPK(int codigoEstablecimiento, int codTitulo, int codSeccion, int codSeccionTexto) {
         this.codigoEstablecimiento = codigoEstablecimiento;
         this.codTitulo = codTitulo;
         this.codSeccion = codSeccion;
         this.codSeccionTexto = codSeccionTexto;
     }
 
-    public short getCodigoEstablecimiento() {
+    public int getCodigoEstablecimiento() {
         return codigoEstablecimiento;
     }
 
-    public void setCodigoEstablecimiento(short codigoEstablecimiento) {
+    public void setCodigoEstablecimiento(int codigoEstablecimiento) {
         this.codigoEstablecimiento = codigoEstablecimiento;
     }
