+++ entity/src/main/java/com/gestor/seguimiento/PlanTituloTextoPK.java
@@ -19,7 +19,7 @@ public class PlanTituloTextoPK implements Serializable {
 
     @Basic(optional = false)
     @Column(name = "codigo_establecimiento")
-    private short codigoEstablecimiento;
+    private int codigoEstablecimiento;
     @Basic(optional = false)
     @Column(name = "cod_titulo")
     private int codTitulo;
@@ -30,17 +30,17 @@ public class PlanTituloTextoPK implements Serializable {
     public PlanTituloTextoPK() {
     }
 
-    public PlanTituloTextoPK(short codigoEstablecimiento, int codTitulo, int codTituloTexto) {
+    public PlanTituloTextoPK(int codigoEstablecimiento, int codTitulo, int codTituloTexto) {
         this.codigoEstablecimiento = codigoEstablecimiento;
         this.codTitulo = codTitulo;
         this.codTituloTexto = codTituloTexto;
     }
 
-    public short getCodigoEstablecimiento() {
+    public int getCodigoEstablecimiento() {
         return codigoEstablecimiento;
     }
 
-    public void setCodigoEstablecimiento(short codigoEstablecimiento) {
+    public void setCodigoEstablecimiento(int codigoEstablecimiento) {
         this.codigoEstablecimiento = codigoEstablecimiento;
     }
